(MacOS) Implemented device_type::read()
Emulate blocking behavior by registering an input report call callback, scheduling the device on the current run loop and then calling CFRunLoopRun(). The callback copies the report into the caller's buffer and then calls CFRunLoopStop(), which has the side effect of returning execution to read(), which then returns to the caller.
Note that scheduling an already scheduled device might cause problems...the documentation doesn't say what happens in that case.